requirements.txt
django-environ==0.9.0
Django==4.1.6
psycopg2==2.9.5
psycopg2-binary==2.9.5
djangorestframework-simplejwt==5.2.2
djangorestframework==3.14.0
djoser==2.1.0
drf-yasg==1.21.5
Pillow==9.4.0